// src/app/admin-settings/admin-route.enum.ts
export enum AdminRoute {
    ORGANIZATION = 'organization',
    DIVISION = 'division',
    APPLICATION_GROUP = 'application-group',
    APPLICATION = 'application',
    SPRINT = 'sprint',
    EMPLOYEE = 'employee',
    VENDOR = 'vendor',
    DESIGNATION = 'designation',
    RELEASE= 'release',
    STATUS = 'status',
    PHASE = 'phase',
    ROLE = 'role',
    REQUIREMENT_TYPE = 'requirementtype',
    ATTRIBUTE = 'attribute'
}